backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-groq
chromadb
sentence-transformers
pydantic
pydantic-settings
python-dotenv
